Problem: code wasn't modular enough to test
Solution: Extract it
